Finished cpp00

+++ ex00/megaphone.cpp
@@ -7,7 +7,7 @@ int main(int argc, char **argv)
 	else
 	{
 		for (int i = 1; i < argc; i++) {
-			for (int j = 0; j < ::strlen(argv[i]); j++)
+			for (size_t j = 0; j < ::strlen(argv[i]); j++)
 			{
 				argv[i][j] = ::toupper(argv[i][j]);
 				std::cout << argv[i][j];
